fix: function definition parsing.
- Fixed the parsing of the function body, where more than one expressions int the function
  were not being parsed.
- To deal with this, a new member 'res' was added to the 'ParsingContext' struct, that keeps
  track of the list of function expressions, so that we can add a new expression at the end
  of the list of expressions, every time we encounter one.
- Added more expressions inside the 'add_int' function under 'example/simple.sy' for testing.
- Modified the 'calculate_file_size' function to return 'size_t' instead of a 'long' to
  fix a comparison between 'long' and 'size_t'.
- Added more warning flags in the Makefile, to get all the warnings in the code.

+++ inc/lexer.h
@@ -40,9 +40,9 @@ typedef long int_cl;
  *
  * @param  file_ptr  [`FILE *`] pointer to the file, for which the
  *                   size needs to be calculated.
- * @return long      The size of the file.
+ * @return size_t    The size of the file.
  */
-long calculate_file_size(FILE *file_ptr);
+size_t calculate_file_size(FILE *file_ptr);
 
 /**
  * @brief Print out the token, pointed to by `curr_token`.
